refactor(Misc): remove ModuleTracker.kt, move totem pop counter to ModuleNotifier.kt
