+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/misc/ModuleNotifier.kt
@@ -25,6 +25,7 @@ import net.ccbluex.liquidbounce.event.events.PacketEvent
 import net.ccbluex.liquidbounce.event.events.PlayerTickEvent
 import net.ccbluex.liquidbounce.event.events.WorldChangeEvent
 import net.ccbluex.liquidbounce.event.handler
+import net.ccbluex.liquidbounce.features.misc.FriendManager
 import net.ccbluex.liquidbounce.features.module.ClientModule
 import net.ccbluex.liquidbounce.features.module.ModuleCategories
 import net.ccbluex.liquidbounce.features.module.modules.misc.antibot.ModuleAntiBot
@@ -34,14 +35,17 @@ import net.ccbluex.liquidbounce.utils.client.regular
 import net.ccbluex.liquidbounce.utils.collection.itemSortedSetOf
 import net.ccbluex.liquidbounce.utils.item.isConsumable
 import net.minecraft.client.player.RemotePlayer
+import net.minecraft.network.protocol.game.ClientboundEntityEventPacket
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoRemovePacket
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoUpdatePacket
 import net.minecraft.world.InteractionHand
+import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.ItemUseAnimation
 import net.minecraft.world.item.Items
 import net.minecraft.world.level.GameType
 import java.util.UUID
+import kotlin.collections.set
 
 /**
  * Notifier module
@@ -77,6 +81,9 @@ object ModuleNotifier : ClientModule("Notifier", ModuleCategories.MISC) {
     private val itemConsumptionCache = Object2ObjectOpenHashMap<UUID, ItemConsumptionState>()
     private val heldItemCache = Object2ObjectOpenHashMap<UUID, HeldItemState>()
     private val observedPlayers = ObjectOpenHashSet<UUID>()
+    private val recentPops = mutableSetOf<UUID>()
+    private val popCounter = mutableMapOf<UUID, Int>()
+    private var lastUpdateTime = 0L
 
     override fun onEnabled() {
         for (entry in network.onlinePlayers) {
@@ -91,6 +98,7 @@ object ModuleNotifier : ClientModule("Notifier", ModuleCategories.MISC) {
         itemConsumptionCache.clear()
         heldItemCache.clear()
         observedPlayers.clear()
+        recentPops.clear()
     }
 
     val packetHandler = handler<PacketEvent> { event ->
@@ -126,6 +134,20 @@ object ModuleNotifier : ClientModule("Notifier", ModuleCategories.MISC) {
                     }
                 }
             }
+
+            is ClientboundEntityEventPacket -> {
+                if (packet.eventId.toInt() == 35) {
+                    val entity = packet.getEntity(world) as? Player ?: return@handler
+                    if (entity == mc.player || FriendManager.isFriend(entity.name.string)) return@handler
+
+                    recentPops.add(entity.uuid)
+                    val pops = popCounter.getOrDefault(entity.uuid, 0) + 1
+                    popCounter[entity.uuid] = pops
+                    lastUpdateTime = System.currentTimeMillis()
+
+                    sendNotifierMessage("${entity.name.string} pop totem $pops times")
+                }
+            }
         }
     }
 
@@ -158,6 +180,12 @@ object ModuleNotifier : ClientModule("Notifier", ModuleCategories.MISC) {
             heldItemCache.keys.retainAll(observedPlayers)
         } else {
             heldItemCache.clear()
+        }
+
+        val currentTime = System.currentTimeMillis()
+        if (currentTime - lastUpdateTime >= 500L) {
+            recentPops.clear()
+            lastUpdateTime = currentTime
         }
     }
